Update moved GoogleImaAd class
